portal-v2/backend/requirements.txt
# Blauweiss Portal Backend
# FastAPI + PostgreSQL + OAuth

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
